2_string_analyzer_service/string_analyzer/string_analyzer_service/serializer.py
from rest_framework import serializers
from .models import StringAnalysisModel


class StringAnalysisSerializer(serializers.ModelSerializer):
    '''
    Serilizer for String Analysis Model
    '''

    class Meta:
        model = StringAnalysisModel
        fields = ['id', 'value', 'created_at', 'properties']
