2_string_analyzer_service/string_analyzer/requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
sqlparse==0.5.5
typing_extensions==4.15.0
